requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.6.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
blinker==1.4
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
colorama==0.4.4; sys_platform == 'win32'
coverage==5.5
dnspython==2.1.0; python_version >= '3.6'
dominate==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.1.2
flask-bootstrap==3.3.7.1
flask-httpauth==4.3.0
flask-login==0.5.0
flask-mail==0.9.1
flask-migrate==2.7.0
flask-sqlalchemy==2.5.1
flask-sslify==0.1.5
flask-wtf==0.14.3
flask==1.1.2
greenlet==1.1.0; python_version >= '3'
httpie==2.4.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==4.0.1; python_version < '3.8'
ipaddress==1.0.23
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jinja2==2.11.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
psycopg2==2.8.6
pygments==2.9.0; python_version >= '3.5'
pysocks==1.7.1
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.17.1
python-editor==1.0.4
requests-toolbelt==0.9.1
requests[socks]==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
typing-extensions==3.10.0.0; python_version < '3.8'
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
visitor==0.1.3
waitress==2.0.0
werkzeug==0.16.0
wtforms[email]==2.3.3
zipp==3.4.1; python_version >= '3.6'
